requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.7.0
click==8.1.6
Django==4.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
flake8==6.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
psycopg2-binary==2.8.6
python-dotenv==1.0.0
django-floppyforms==1.9.0
django-leaflet==0.29.0
django-geojson==4.0.0
django-leaflet==0.29.0
pygeoif==1.2.0
